Improvement: AWS Workflow Video on demand terraform module
